requirements.txt
pandas
mlflow == 2.10.0
notebook
numpy
matplotlib
python-box == 6.0.2
pyYAML
tqdm
ensure == 1.0.2
joblib
types-PyYAML
Flask
Flask-Cors
pathlib
os
pandas
numpy
typing
dagshub
matplotlib
seaborn
scikit-learn
# pip install python-box
-e . 